minor tweak to color coding
This is one of the most helpful images I know of when learning these amazing modules. I downloaded it and used it in the computerscare Custom Blank for reference. I had to download the image a few times. Each time, I'd tweak it by adding a 5px border. This allows the image to sit in the blank and look correct on the face of the module. After modifying the image once more, I decided to propose my minor update so that people I refer to the documentation can more easily move the image into their rack